symlinks.h: move declarations for symlinks.c functions from cache.h
